requirements.txt
Flask>=3.0.2
flask-cors>=4.0.0
tensorflow==2.15.0
opencv-python-headless==4.8.0.76
numpy>=1.25.2
matplotlib==3.7.1
pillow==9.4.0
